Fixed a invalid publish

+++ server/src/services/publishSchema.ts
@@ -9,8 +9,10 @@ const SCHEMA_NAME = 'schema/epicast-demoserver-feed1-schema.yaml'
 const SCHEMA_TEMPLATE_PATH = './src/public/epicast-demoserver-feed1-schema.yaml'
 
 export async function publishSchema(feed: Feed, log: FeedLog): Promise<void> {
-  logger.info("publishing schema")
-  const rawSchema = readFileSync(SCHEMA_TEMPLATE_PATH, { encoding: 'utf8' })
-  await feed.putObject(SCHEMA_NAME, rawSchema)
-  log.update(SCHEMA_NAME)
+  if (!await feed.doesObjectExist(SCHEMA_NAME)) {
+    logger.info("publishing schema")
+    const rawSchema = readFileSync(SCHEMA_TEMPLATE_PATH, { encoding: 'utf8' })
+    await feed.putObject(SCHEMA_NAME, rawSchema)
+    log.add(SCHEMA_NAME)
+  }
 }
